+++ docs/content/docs/uninstallation_guide.md
@@ -36,6 +36,7 @@ The following sections detail which Kubernetes resources are cleaned up and whic
 1. Existing Envoy sidecar containers
     - Redeploy application pods to delete sidecars
 1. Envoy bootstrap config secrets (stored in the application namespace)
+1. If WebAssembly stats were enabled, a ConfigMap named `envoy-stats-wasm` in each application namespace
 1. Namespace annotations, including but not limited to `openservicemesh.io/monitored-by`
 1. Custom resource definitions ([CRDs](https://github.com/openservicemesh/osm/tree/main/charts/osm/crds))
 1. Vault resources provided by the user
